Refactor test_json_descriptor
